JCBC-1066: Add support for FTS Geo & TermRange Queries & Advanced Sort.
Motivation
----------
Couchbase Server 5.0 ships with some new FTS capabilities which need to
be exposed properly in the SDK these are:

 - Geo Bounding Box Query
 - Geo Distance Query
 - Term Range Query
 - Advanced Sorting (Field, GeoDistance, Id, Score)

Modifications
-------------
This changeset adds the three new query types as well as the new sorting
capabilities exposed as a simple builder, very similar to the facets API.

Unit tests have been added to make sure the new capabilities produce
the query as intended.

In addition, the sort param also allows for a "JsonObject" to be passed in,
this allows for forwards compatibility when a new sort mechanism is added
in future server versions, the user can always fall back to that type
and still benefit from querying.

Result
------
Support for the new querying mechanisms as well as advanced sorting
capabilities.

Change-Id: I8f4caca1629925bc0caaacb82f39306aace87bf2
Reviewed-on: http://review.couchbase.org/76246
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>