JCBC-1072: Add covenient authenticate overload.
Motivation
----------
Most of the time when user-based authentication is used one just
needs to supply a username and a password. Right now this is only
possible by doing:

 cluster.authenticate(new PasswordAuthenticator("user", "pw"));

This adds some friction to the user experience since the user
first needs to figure out what an authenticator is, which instance
one should be using and so forth.

Modifications
-------------
This changeset adds a simple overload which should suffice most
of the time and will allow the user just to authenticate with
a username and a password string like this:

 cluster.authenticate("user", "pw");

The old overload is still available for advanced use cases of
course.

Result
------
Much easier way to authenticate against a cluster with user
credentials.

Change-Id: Icb8aa20fd5f3da44436dd5e8745ebdb95fe11173
Reviewed-on: http://review.couchbase.org/76637
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>