JCBC-1076: Add support for "Enhanced Error Messages"
Motivation
----------
In combination with the corresponding core-io changes, this
changeset adds support for "enhanced error messages", which
allows the server to send more detailed information on
errors if available.

Modifications
-------------
The CouchbaseAsyncBucket code is modified to convert error
messages into "enhanced" ones and store it inside the message
of the exception if the enhanced msg is not null.

Result
------
Support for enhanced error messages.

Change-Id: I7b687c959787790e7661b627d11a155f83cbda17
Reviewed-on: http://review.couchbase.org/79483
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Subhashni Balakrishnan <b.subhashni@gmail.com>