JCBC-1110: Add Experimental TransparentReplicaGetHelper
Motivation
----------
Users sometimes need assistance with transparently failing
over to a replica get when the primary get fails. This little
helper class provides utility methods to help with exactly that.

Modifications
-------------
This changeset introduces a helper class which first performs
a regular kv get and if that one fails it fails over to a
replica get and grabs the first result it gets back. This is
the most common situation users are using both commands together.

Result
------
Transparent get followed potentially by replica get is a common
use case and now covered by a simple helper class.

Change-Id: Ib1aa0f0276147977fe81bcfefb92f80f7906cc14
Reviewed-on: http://review.couchbase.org/80443
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: David Nault <david.nault@couchbase.com>