JCBC-1105: Make custom param settings deterministic in order.
Motivation
----------
Both the DefaultBucketSettings and the N1qlParams have HashMaps
which allow to store "custom" params which in turn get stored
in a JSON array on the wire. While the order for them ultimately
don't matter, there is a certain expectation that the insert
order is preserved, and it also helps with assertability
in unit tests.

Modifications
-------------
The HashMap instances are replaced with LinkedHashMap instances
that preserve the insertion order. This also resulted in small
changes for the N1qlParamsTest which now has unit tests (correctly)
failing on OSX (they always did on linux as CI showed) and they
are now corrected.

Result
------
Deterministic behavior based on insertion order for the N1ql custom
params as well as the bucket settings.

Change-Id: I8a8f33b4bef53eb9f704c462dfb3a93cfb63f22d
Reviewed-on: http://review.couchbase.org/82241
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>