JCBC-1082: Implement support for Subdoc GetCount
Motivation
----------
In Couchbase Server 5.0, a new lookup command (get count) is introduced
which allows to fetch the number of items for a given path in a document.

Modifications
-------------
This changeset adds support for the lookup in builder (and async version)
so that they support taking getCount as arguments. Tests are adapted to
test both the single and multi case variants.

Result
------
java-client support for subdocument get count.

Change-Id: I8383a3017fe8cbd21d1b330a44c5020c5cfd61b1
Reviewed-on: http://review.couchbase.org/82444
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Build Bot <build@couchbase.com>