JCBC-1115: Enforce Xattr first ordering on subdoc builders.
Motivation
----------
When using the subdoc xattr API, a server side constraint is that
the xattr-based commands always need to come before the non-xattr
ones.

This should be enforced on the client side to give rapid feedback
and not wait for cryptic errors at runtime.

Modifications
-------------
This changeset introduces a new subdoc exception and modifies both
the lookup and the mutation builder to enforce the constraint
at execute() time.

Result
------
The server-side requirement is enforced at the client side, leading
to rapid feedback.

Change-Id: I96568b7f4ff952deb39a231b1d0625959f2a0ff4
Reviewed-on: http://review.couchbase.org/82453
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>