JCBC-1100: Document that expiry on counter only works on creation.
Motivation
----------
An edge case happens with counters and expiry in that the expiry
is only honored when the document is created, not updated. This
differs as with the other commands  but is currently a limitation
on the server.

Modifications
-------------
To clarify, all affected javadoc blocks have been updated with
a note.

Result
------
Less confusion when reading the javadoc about this edge case.

Change-Id: Ia898d9070c1082e9d76f3a6db673c9dbff1935e1
Reviewed-on: http://review.couchbase.org/82498
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: David Nault <david.nault@couchbase.com>