JCBC-1117: Add test case for CoreSendHook
Motivation
----------
core-io added the send hook but a test case to both document and
make sure no regressions are introduced should be added.

Modifications
-------------
As part of adding the test is was discovered that the artificial
limitation added to deferAndWatch (only work on subjects) needs
to be lifted because chained in observables are not subjects
anymore at this point, but it has been stable enough to do just
that confidently.

The test adds a hook which records an upsert operation and its
latency, asserting on the result and also logging it, acting
as an example.

Result
------
Adding a core hook and recording operations as a regression test
as well as sample code.

Change-Id: I19c04696c1b217cac10089c3b0b91b2768e0ebb1
Reviewed-on: http://review.couchbase.org/82626
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>