JCBC-1125: Add support for the basic healthCheck API.
Motivation
----------
Building on the changes from core-io, the java-client needs to
expose a simple health check API to quickly gather important
information about the underlying endpoints/sockets.

Modifications
-------------
This changeset uses the new requests and responses to query
core-io for the services state and then exposes it via the
healthCheck() method on Cluster/AsyncCluster appropriately.

Result
------
The healthCheck() API is available and marked as experimental.

Change-Id: Ibfd15a7b3cc786b2a162532fe93ca4535ddb5a74
Reviewed-on: http://review.couchbase.org/83081
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>