JCBC-1141: Add ByteArrayDocument & Transcoder
Motivation
----------
In the past, the canonical document type users were falling
back to when dealing with raw byte buffers is the BinaryDocument.
Because this document type puts the burden of refcounting on the
user, lots of edges can come up which need to be dealt with on
a case-by-case basis.

Modifications
-------------
This changeset introduces a new document & transcoder which
operates on a byte[] which, at the cost of a memcopy, allows
users to deal with raw data but don't have to care about
refcounting and edge cases. It has been useful in the past
in specific deployments, so it makes sense to provide it to
the broader public as well.

Result
------
A new document type which should be used in general over the
very specific BinaryDocument type.

Change-Id: Idaab0429d40774c3bc67eac294ab4e9c6f0c0a3c
Reviewed-on: http://review.couchbase.org/84198
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>