JCBC-1142: Rename createParents to createPath
Motivation
----------
As part of the XATTR/Subdocument RFC it was decided
to unify the "createParents" attribute to "createPath"
which better reflects the user intent.

Modifications
-------------
This changeset applies this change in a fully backwards-
compatible manner. API has been renamed where either internal
or just using boolean flags. In the other places where actually
method changes are needed, the appropriate methods have been
marked as deprecated and the right proper methods implemented
in addition.

Result
------
The Java SDK now properly exposes "createPath" instead of
"createParents", but its still available for backwards
compatibility and will be only removed with the next major
SDK bump.

Change-Id: I966ae0fb940295004bce94d5a02a5458fd2c889b
Reviewed-on: http://review.couchbase.org/84749
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>