JCBC-1147: Add health check ping command support.
Motivation
----------
In addition to the refactored diagnostics command,
the RFC calls for a bucket-level "service ping"
facility which has similar output to the diagnostics
but proactively calls the services on their endpoints.

Modifications
-------------
This changeset simply exposes the methods with their overloads
on the Bucket and AsyncBucket and forwards it to the HealthPinger.

Result
------
ping support on the bucket!

Change-Id: I5d1dab046626977184e390c0d11d9770744c9e27
Reviewed-on: http://review.couchbase.org/87094
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>