JCBC-1163: Log value on n1ql row decode failure.
Motivation
----------
If, for whatever reason, a n1ql row fails to decode
with jackson we should log the original value as part
of the exception message so after the fact it is possible
to figure out what went wrong.

Modifications
-------------
The exception message has been adapted to include the original
value as a utf8 string and a test has been added to make sure
this is actually the case.

Result
------
The exception now has the original value as part of its message.

Change-Id: I995afeac88456050f5b48c8c94a478b6c7a6eed7
Reviewed-on: http://review.couchbase.org/88180
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>