JCBC-1164: Add Tracing to Get & Exists.
This changeset adds tracing and timeout functionality to the
following requests:

 - get
 - getAndTouch
 - getAndLock
 - exists

Since the bucket impl is growing, its time to refactor the
actual implementations into their own helper classes.

Modifications
-------------
Other than refactoring the code it has been changed to perform
timeouts through the timeout() operator, adds async APIs with
timeout overloads and if a timeout exception is raised takes
it, and rethrows it with additonal information based on the
tracing RFC.

Result
------
Get, exists and variants are now working with respect to the
tracing rfc.

Change-Id: I8758d9bd92a8abc4928e13a7fc32dbd7b7e729db
Reviewed-on: http://review.couchbase.org/89951
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>