JCBC-1164: Add Tracing to getFromReplica
This changeset adds tracing to getFromReplica. It creates
an internal parent span for each execution. The threshold
log will only pick up the individual get or replica get
requests, since the overall parent span has no real benefit
in this case (but it does when used in a distributed tracing
context).

Change-Id: I5eb800fc6099a7b523ac0b1ffd65cd8dd7edf112
Reviewed-on: http://review.couchbase.org/89962
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>