JCBC-1203: Add CertAuthenticator and related checks
Motivation
----------
This changeset brings the Java SDK API-wise in line
with the SDK-RFC around client certificate authentication.
It worked before but this change brings in the CertAuthenticator
as well as adds all kinds of sanity checks so that invalid
auth combinations are rejected.

Modifications
-------------
This changeset adds the CertAuthenticator (which can be used
as a singleton) and then adds checks to the CouchbaseAsyncCluster
so that invalid combinations depending on the environment settings
are not allowed and error quickly.

Tests have been added to verify those in a unit-test setting.

Result
------
The Java SDK is in line with the RFC API requirements and error
handling sections.

Change-Id: I35202e466a69b0d3d29d913c3b74d3b5558edea0
Reviewed-on: http://review.couchbase.org/94200
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>