JCBC-1158: Support annotations in parent classes
Motivation
----------
On our experimental repository API, it was reported that if the
ID (or any other field for that matter) was stored on a parent
class it is not picked up properly by our mapper.

Modifications
-------------
This changeset makes sure that all the fields looked at are
recursively picked up from parent classes as well so that
if a child class is passed in all the proper fields from
parent classes can be used.

Test case has been added to verify this.

Result
------
Support for fields with annotations in parent classes.

Change-Id: I219e1733476d2958699ddec8b4b52eacc6b7f19b
Reviewed-on: http://review.couchbase.org/94676
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-on: http://review.couchbase.org/95134