JCBC-1209: Audit and add explicit subscribers on req creation
Motivation
----------
In an earlier changeset the explicit subscriber management on
the request has been added, but some of the APIs have been
overlooked - this can lead to inconsistent behavior
in core-io.

Modifications
-------------
This changeset audits all calls to core.send(...) and where
needed adds the explicit subscriber management on the request
since core-io checks if a subscriber is attached and if not
it never treats it as non-active. Especially n1ql requests
in this regard have been treaded differently before this
changeset compared to kv requests.

Result
------
Complete audit of .send() APIs and every request now properly
handles it subscriber and can be checked against timeouts in
core-io.

Change-Id: I4d85934565ec89390232b2f0d21751d59a291b4a
Reviewed-on: http://review.couchbase.org/95189
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>