JCBC-1209: Audit and add explicit subscribers on req creation
Motivation
----------
In an earlier changeset the explicit subscriber management on
the request has been added, but some of the APIs have been
overlooked - this can lead to inconsistent behavior
in core-io.

Modifications
-------------
This changeset audits all calls to core.send(...) and where
needed adds the explicit subscriber management on the request
since core-io checks if a subscriber is attached and if not
it never treats it as non-active. Especially n1ql requests
in this regard have been treaded differently before this
changeset compared to kv requests.

Result
------
Complete audit of .send() APIs and every request now properly
handles it subscriber and can be checked against timeouts in
core-io.

Change-Id: I71f752384af7ac4fa88534531401901f92a1fb24
Reviewed-on: http://review.couchbase.org/95227
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>