JCBC-1186: Pick up and allow to configure analytics timeout
Motivation
----------
If no explicit server side timeout is provided by the user,
similar to the n1ql API, an implicit server side timeout
should be set which aligns with the client side timeout to
not let them run to infitiny if the client is not interested
anymore.

Note that since the server only recently introduced the
"timeout" config option, this change is not backwards
compatible with older pre-release analytics server
versions if they reject unknown config params!

Modifications
-------------
This changeset works by making two changes:

 - first, a new serverSideTimeout config option is
   introduced which allows to explicitly set the
   server side timeout.
 - second, the bucket code checks if an explicit
   timeout if set and if this is not the case sets
   it to the client side timeout.

Result
------
If the client stops listening, this will make sure
that the server side also stops processing the
request.

Change-Id: I033ef31a1545b7b1f9ee94f9a1234db94c2a9a22
Reviewed-on: http://review.couchbase.org/95355
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>