JCBC-1213: Search Query sorting order needs to be "desc"
Motivation
----------
A bug was reported by the community that field sort order
"descending" was wrong when sent to the server, and indeed
it needs to be "desc" on the wire.

Modifications
-------------
The SearchSort serialization key has been adapted to reflect
the correct encoded key and the tests have been adapted
as well.

Result
------
It is now possible to set the descending field sort order,
since previously the server would ignore it.

Change-Id: I09318e777f20b391ebbb4d9ae264f2b36df87685
Reviewed-on: http://review.couchbase.org/95993
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>