JCBC-1221: Implicitly propagate parent span if available
Motivation
----------
Right now it is not possible to specify a parent span for the
operations traced inside the SDK. This is unfortunate since
context is lost which should be propagated if possible.

Modifications
-------------
This changeset latches onto the fact that there is an active
span in scope (mostly implemented through thread locals in
the actual impls) and uses this one explicitly if provided.

This allows to run code like this to automatically propagate:

try (Scope scope = tracer.buildSpan("my-op").startActive(true)) {
  bucket.get("document");
}

Since most of the time thread locals are used, when used with callbacks
the scope needs to be activated explicitly in callbacks:

bucket.async().get("airline_10123").flatMap((Func1<JsonDocument, Observable<?>>) jsonDocument -> {
      tracer.scopeManager().activate(scope.span(), false);
      return bucket.async().upsert(jsonDocument);
}).toBlocking().single();

Also added an environment switch to explicitly disable the
propagation, but it's enabled by default.

Result
------
Automatic parent span propagation into the SDK scope, can be disabled
if needed.

Change-Id: Id89e12d3ec2e0cadc1df3b1ee059b8c0817a03c0
Reviewed-on: http://review.couchbase.org/96766
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>