JCBC-1223: Do not add timeout if it's zero.
Motivation
----------
A bug has been found which causes operations to time out very
quickly under the following conditions:

 - The Async API is used
 - No explicit timeout is provided
 - Durability requirements are specified (PersistTo/ReplicateTo)
 - The operation is one of the following KV operations:
    * insert, upsert, replace
    * remove
    * counter
    * append, prepend

This manifests mostly when calling a method from above like this:

  bucket.async().upsert(doc, PersistTo.MASTER)...;

This will internally schedule a timeout with 0, meaning it will
time out right away. Since all other combinations actually specify
a timeout they are not affected.

Modifications
-------------
This changeset audits the java-client codebase for cases where a
possible 0 timeout can be used and adds integration tests for the
KV cases outlined above.

Not applying a timeout if it is 0 is the backwards compatible way
since in previous releases also no timeout was in place for these
cases.

Result
------
Backwards compatible behavior (no timeout) for async KV operations
with durability requirements. Only 2.6.0 is affected, previous
2.5.x releases are not.

Change-Id: I205927704ca606e98dd4e610b1658c71a1702204
Reviewed-on: http://review.couchbase.org/96843
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>