JCBC-1226: Initialize AnalyticsParams if not provided.
Motivation
----------
If no analtics params are provided by the user, the default
implementation should be used so that downstream code does
not result in a NPE.

Modifications
-------------
Similar to what N1QL does, the query initializes the analytics
params now to the default object if not provided so that there
is no NPE down the code paths.

Result
------
No NPE when bucket.query(AnalyticsQuery.simple("...")); is used.

Change-Id: Idfe12d02fdc09a6b896ae1c882e2712f592d4e50
Reviewed-on: http://review.couchbase.org/97384
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>