JCBC-1236: Add support for analytics priority param.
Motivation
----------
This changeset exposes the priority setting as a param and
passes it down to the core-io layer.

Modifications
-------------
Note that according to the RFC it is exposed as a boolea for now,
but since the server support it as an integer we are storing it
like that internally. This makes it trivial to open it up further
in the future.

Change-Id: I24bd78cd2f1be450cf364dd9c9b1d745d7741b5f
Reviewed-on: http://review.couchbase.org/98375
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>