JCBC-1229: Add support for parameterized analytics queries.
Motivation
----------
This changeset adds support for parameterized queries, very similar
to what we already expose for N1QL.

Modifications
-------------
ParameterizedAnalyticsQuery has been added which extends from
the simple one and works the same but it also injects either
positional or named arguments.

Unit tests added to make sure the params are passed in the
right way.

Change-Id: I2d42a5d8fb81b18e5e62faceed43832c897650ae
Reviewed-on: http://review.couchbase.org/98435
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>