JCBC-1245: Better defaults for BucketSettings
Motivation
----------
Previously a user had to always specify a quota, but setting
it to 100mb by default is a reasonable setting. Also, the
password is not required anymore with RBAC.

Together, these changes make it easier to get started with the
Bucket Settings and improve the out-of-the-box experience.

Tests have been added.

Change-Id: I6e2baac19a741623569f659e323fd4690319fe4d
Reviewed-on: http://review.couchbase.org/99933
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>