JCBC-1232: Transparent retry for search http 429
Motivation
----------
Since Couchbase Server 6.0 the search engine respects the memory
quota in a hard way and will push back to the client if it is
overloaded.

To deal with this, a code change is in place that applies a retry
logic but will also propagate the error if it can't work it out
after the retries are exhausted.

For now this is only configurable through system properties, but
aligned with the other SDKs. The current retry delay
intervals might change over time.

Test have been added to assert the functionality.

Change-Id: I5177a5b2642e0a056833691ceed57b6f42b7fc27
Reviewed-on: http://review.couchbase.org/99952
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>