JCBC-1253: Support subdoc expansion of macro
Motivation
----------
Adding server-side expansion of xattrs macros such as "${Mutation.CAS}"
This is required for the multi-document atomicity project.
As this is advanced functionality that we may not want users to access,
the elements added to the API are marked private.

Modifications
-------------
Added expandMacros fields to all required builders, and pass down a new
0x10 flag to the server to indicate that macros should be expanded for
a particular field.

Result
------
Tested that ${Mutation.CAS} is replaced with the doc's CAS as expected.

Change-Id: I0443578adbe57dae960389ddd9c7effce524ad15
Reviewed-on: http://review.couchbase.org/101059
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Graham Pople <grahampople@gmail.com>