JCBC-1259: Expose AsyncCluster from Cluster
Motivation
----------
The Bucket API exposes the AsyncBucket through a call to async(),
but the Cluster API does not do the same. There are situation where
(for example to grab diagnostics in an async context) this is
useful.

Modifications
-------------
Cluster now exposes the async variant through the async() method,
similar to Bucket and other likewise interfaces.

Result
------
The AsyncCluster can now be accessed through a Cluster reference.

Change-Id: Id18d9c43c8983e2b7f733f716dc27aa55d7e6d87
Reviewed-on: http://review.couchbase.org/100690
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>