Avoid premature DNS lookups
The SDK should be able to bootstrap the list of the seed nodes not only
from IP addresses, but also from DNS names. For example, if a Couchbase
cluster is managed by Kubernetes, its nodes don't have fixed IP
addresses: they are exposed by hostnames. Hence, we want to make
sure that the SDK will be able to pick up any changes of the IP address
associated with the hostname. Currently, it's not possible, because SDK
prematurely resolves DNS names to IP addresses during its startup and
seeds the client with IP addresses. As a result, the SDK is not resilent
to changes in the server, e.g. we can't restart it without restaring all
clients.