JCBC-1266: Allow to compile against Java 11
Motivation
----------
To make progress with newer java versions, we need to make sure the client
builds and runs fine against those. A couple changes were required to do
so.

Modifications
-------------
Like with core-io, the major trouble was javadoc. We had to get rid of the
pegdown doclet plugin and then javadoc started to throw a tantrum on certain
javadocs, even with doclint off.

This changeset removes all those warnings and allows for a green build
process.

Result
------
the java client now compiles against java 11

Change-Id: I8b2b1e4a31d1fbb841cdbb123599d1cbd98b8fc5
Reviewed-on: http://review.couchbase.org/101587
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>