Prepare java-client to run in CI under the mock.
Note that this changeset touches a lot of files which do not work correctly
under the mock either because the code is too tied towards the actual cluster
management or the mock just plainly does not support it.

We still run lots of tests though, so I deemed its better to run those and
ignored the rest, which we can revisit test by test and rework to reenable
if needed.

Change-Id: I10bb980ccfbbf6d971961835f27664ffb496b4e5
Reviewed-on: http://review.couchbase.org/101905
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>