JCBC-1287: Feed query statement into core for tracing tag
Motivation
----------
core-io added support for adding a standard opentracing tag into
the span for n1ql queries, so this changeset follows suit to feed
the raw statement into it.

Modifications
-------------
The executor now passes the statement in string form down to core-io
which picks it up and adds it as a span tag if needed.

Change-Id: Ibbdfa23178b9680e7dbda468548cf063257ce623
Reviewed-on: http://review.couchbase.org/104189
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: David Nault <david.nault@couchbase.com>