JCBC-1323 - Handle empty FTS error block
Motivation
----------
Sometimes the Search service can return a response with a status
block similar to the following:

```
{
  "total": 6,
  "failed": 0,
  "successful": 6,
  "errors": {}
}
```

The Java SDK assumes that if the "errors" field is present then there
must be FTS errors to propagate to the application, when in fact in
this specific instance there are actually no errors.
The result of this is that users applications error out when there
haven't been any issues.

While this is a server-side inconsistency (covered in MB-33830), the
Java SDK should be more defensive to this situation.

Modifications
-------------
This commit adds a check to see if the "errors" field is empty before
creating the error observable.

If the "errors" block is empty then it will now return an empty
observable instead of an error observable.

Result
------
Now the SDK will not throw an exception if it receives a response
containing an empty error block.

Change-Id: I8c2bb691178280d729fe322abf7c527522fedb7a
Reviewed-on: http://review.couchbase.org/107927
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Graham Pople <grahampople@gmail.com>