JCBC-1333 Reflection triggers NoClassDefFoundError for CryptoManager
Modifications
=============
Invert the dependency relationship between the client and the encryption
library. The classes that define the API are now in the client.
For now this means they will be duplicated by the encryption library.
In the future we can release a new version of the encryption library
containing only the provider and keystore implementations.

Change-Id: I91264e365e1849ae2ebd730b5914d588e3e158ad
Reviewed-on: http://review.couchbase.org/108467
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: David Nault <david.nault@couchbase.com>