Add script for running integration tests under CI
Motivation
==========
The "runtest" Python script is used by Jenkins to set up the
integration test environment and run the tests. Previously it was
an external gist on a personal developer account; it's more at home
here in the source tree.

Modifications
=============
Add the script under `src/integration/bin`.

Update .gitignore to not ignore "bin" directories... or Ivy jars, which
probably haven't been used in many years.

Change-Id: Iccf697c285ab1b806a2d257beaa786a2ebe513b6
Reviewed-on: http://review.couchbase.org/109214
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: David Nault <david.nault@couchbase.com>