Eliminate ram-quota and storage-mode flags
This is just an experiment - but it matches successful usage I've
seen in the go integration tests.  Figure it is a place to start
given that our tests die (with timeouts) at different points in
the tests.  So - seems like it must be something with how we
provision the integration test cluster.

Change-Id: I34bf037dda4683e9ea0a92c6aaee254f3026b063
Reviewed-on: http://review.couchbase.org/111492
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>