JCBC-1366 CouchbaseMock query ping times out
Motivation
==========
The DiagnosticTests.shouldRunPing fails consistently when mock is enabled.
Since jenkins runs with the mock, this needs to be fixed.  JVMCBC-687 was
fixed, which corrected the behaviour against the real server, however our
mock still has the issue.  Lets first work around this by not pinging the
mock's Query service.  Later, we can fix the mock (and un-do this change).
Also worth examining the api's logic here - we get a response back, just
not what we expect.  But the api waits till the timeout occurs before
returning as a timeout.  Perhaps we can do something else?

Modification
============
The test just avoids the bug in the CouchbaseMock for now, by specifying
the specific services to ping, when the mock is enabled.

Change-Id: Ia4b49348a8962b406b44127a71b3e303f257a967
Reviewed-on: http://review.couchbase.org/111512
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>