JCBC-1366 CouchbaseMock query ping times out
Motivation
==========
The previous fix for this issue didn't really fix the -DuseMock
not being honored properly.  Seems it still wasn't, and this needs
to be addressed as part of that fix

Modification
============
We need the pom to properly handle test resources, and the filtering
for them.  Also, we needed a small change - the original leading backslash - to
properly load the properties file in the first place.

Change-Id: I8cf21824b0d4a5f8b369ca67bcf44a1fb62c87fb
Reviewed-on: http://review.couchbase.org/111755
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>