JCBC-1373 Fix integration tests
Motivation
==========
Getting Jenkins integration tests to pass, like they do when run
locally.

Modification
============
We will ignore the BucketManagerIndexManagementTests just temporarily.
It appears the cluster we spin up experiences issues (maybe crashes) after
this test is run.  To verify,  we can ignore it for now.  Once we
figure out if it does/doesn't actually cause the issue, then we can
re-enable this test.

Change-Id: Iae7b37c9b9c7392ed3d530a1049b5e70fa7ada68
Reviewed-on: http://review.couchbase.org/111918
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>