JCBC-1373 Fix integration tests
Motivation
==========
Getting Jenkins integration tests to pass, like they do when run
locally.

Modification
============
The tests fail, typically, trying to open a bucket at the beginning
of a test.  But when I look, that bucket exists, and all the services
are ok.  So - rather that reuse the cached bucket, lets be sure when
we call openBucket we really reopen it.

Also - the ad_hoc buckets were being named 2 times, which makes very
long bucket names and is just sorta confusing to look at.  Easy enough
to stop that.

Change-Id: I6eaa96a719b41fd3a909dc69a1ecdf15617dacc2
Reviewed-on: http://review.couchbase.org/112108
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>