JCBC-1373 Fix integration tests
Motivation
==========
Getting Jenkins integration tests to pass, like they do when run
locally.

Modification
============
ClusterDependentTest was fixed, however a number of tests that
didn't use it were not shutting down environments, which lead to
a considerable number of warnings when running integration tests.
Worth eliminating, so this attempt to do exactly that.  Will do
same for the core tests shortly.

Change-Id: Ic51c91282f6639b3830a4b0620e23bcf409e62d8
Reviewed-on: http://review.couchbase.org/112181
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>