JCBC-1373 Fix integration tests
Motivation
==========
Getting Jenkins integration tests to pass, like they do when run
locally.

Modification
============
This will rerun a test if it fails, when executing the integration
tests on Jenkins.  There are a small number of tests which are somewhat
sensitive to timing, and fail maybe 1 in n times on jenkins.  This is
a simple workaround, perhaps.

Change-Id: I331cfc6bfb392bd3c8c60eef2e320013f0fb5693
Reviewed-on: http://review.couchbase.org/112556
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: David Kelly <davidmichaelkelly@gmail.com>