JVMCBC-713: Do not try to resolve in ConnectionString
Motivation
----------
Not only do we need to avoid resolving early in the core, also for
the actual bootstrap list nodes we need to make sure that they are
not being resolved. With this change the hostname or ips are passed
through 1:1 as they are provided, making sure that the seed host
list is deterministic and can be debugged properly in the logs.

Change-Id: I5bf004867238cf10d44f17d99a501d401f8b5516
Reviewed-on: http://review.couchbase.org/112562
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>